requirements.txt
# add packages to be installed here
pymongo
python-dotenv